fix: drop the redundant Dev conversion that fails unconvert on linux
stat.Dev is already uint64 on Linux, so converting it there is a no-op that
golangci-lint's unconvert rejects. The file is build-tagged !windows, so
this never showed up in a lint run on a Windows checkout - only in CI.

The conversion is dropped rather than adjusted, because Dev is not the same
integer type across unix platforms: uint64 on Linux, int32 on darwin.
Formatting it with %d handles either. The value is only ever compared for
equality when grouping filesystems, so any stable per-device rendering is
equivalent.

+++ internal/fsutil/filesystem_id_unix.go
@@ -21,5 +21,8 @@ func getFilesystemID(path string) (FilesystemID, error) {
 		return "", fmt.Errorf("failed to read device ID for %s", path)
 	}
 
-	return FilesystemID(fmt.Sprintf("dev:%d", uint64(stat.Dev))), nil
+	// Dev is not the same integer type across unix platforms - uint64 on Linux,
+	// int32 on darwin - so it is formatted rather than converted. The ID is only
+	// ever compared for equality, so any stable per-device rendering will do.
+	return FilesystemID(fmt.Sprintf("dev:%d", stat.Dev)), nil
 }
